Исправления по баг-листу

+++ source/js/modules/form.js
@@ -2,10 +2,23 @@ const form = document.querySelector('.booking__form-wrapper').querySelector('for
 const inputName = document.querySelector('#name');
 const inputPhone = document.querySelector('#phone');
 const inputEmail = document.querySelector('#email');
+const formErrorMessage = form.querySelector('.booking__message');
+const re = /^[\+]?[0-9]*$/;
 
-form.addEventListener('submit', () => {
-  localStorage.clear();
-  localStorage.setItem('name', inputName.value);
-  localStorage.setItem('phone', inputPhone.value);
-  localStorage.setItem('email', inputEmail.value);
+inputPhone.addEventListener('focus', () => {
+  formErrorMessage.classList.remove('booking__message--error');
+});
+
+form.addEventListener('submit', (evt) => {
+  evt.preventDefault();
+
+  if (!re.test(inputPhone.value)) {
+    formErrorMessage.classList.toggle('booking__message--error');
+  } else {
+    localStorage.clear();
+    localStorage.setItem('name', inputName.value);
+    localStorage.setItem('phone', inputPhone.value);
+    localStorage.setItem('email', inputEmail.value);
+    evt.target.submit();
+  }
 });
